+++ core/src/com/metaminers/game/objects/buildings/TowerBunker.java
@@ -9,7 +9,7 @@ import com.badlogic.gdx.graphics.Texture;
 public class TowerBunker extends AbstractBuilding {
     private static final int price = 30;
     private static final int damage = 10;
-    public TowerBunker(float posX, float posY) {
+    public TowerBunker(int posX, int posY) {
         this.hp = 250;
         this.texture = new Texture(Gdx.files.internal("buildings/tower3.png"));
         this.posX = posX;
