+++ core/src/com/metaminers/game/objects/buildings/TowerDouble.java
@@ -9,7 +9,7 @@ import com.badlogic.gdx.graphics.Texture;
 public class TowerDouble extends AbstractBuilding {
     private static final int price = 40;
     private static final int damage = 40;
-    public TowerDouble(float posX, float posY) {
+    public TowerDouble(int posX, int posY) {
         this.hp = 100;
         this.texture = new Texture(Gdx.files.internal("buildings/tower4.png"));
         this.posX = posX;
